Move fake detector config to instrument files
